# docs/swagger.yaml
basePath: /
definitions:
  main.CPU:
    properties:
      guest:
        type: number
      guest_nice:
        type: number
      idle:
        type: number
      iowait:
        type: number
      irq:
        type: number
      nice:
        type: number
      softirq:
        type: number
      steal:
        type: number
      system:
        type: number
      user:
        type: number
    type: object
  main.CreateResponse:
    properties:
      id:
        type: string
      state:
        type: string
    type: object
  main.DiskStat:
    properties:
      io_in_progress:
        type: integer
      name:
        type: string
      reads_completed:
        type: integer
      reads_merged:
        type: integer
      sectors_read:
        type: integer
      sectors_written:
        type: integer
      time_io:
        type: integer
      time_io_weighted:
        type: integer
      time_reading:
        type: integer
      time_writing:
        type: integer
      writes_completed:
        type: integer
      writes_merged:
        type: integer
    type: object
  main.ExecCommand:
    properties:
      cmd:
        items:
          type: string
        type: array
    type: object
  main.ExecResponse:
    properties:
      output:
        type: string
    type: object
  main.FileFD:
    properties:
      allocated:
        type: integer
      maximum:
        type: integer
    type: object
  main.Memory:
    properties:
      active:
        type: integer
      buffers:
        type: integer
      cached:
        type: integer
      dirty:
        type: integer
      inactive:
        type: integer
      mem_available:
        type: integer
      mem_free:
        type: integer
      mem_total:
        type: integer
      shmem:
        type: integer
      slab:
        type: integer
      swap_cached:
        type: integer
      swap_free:
        type: integer
      swap_total:
        type: integer
      vmalloc_chunk:
        type: integer
      vmalloc_total:
        type: integer
      vmalloc_used:
        type: integer
      writeback:
        type: integer
    type: object
  main.NetworkDevice:
    properties:
      name:
        type: string
      recv_bytes:
        type: integer
      recv_compressed:
        type: integer
      recv_drop:
        type: integer
      recv_errs:
        type: integer
      recv_fifo:
        type: integer
      recv_frame:
        type: integer
      recv_multicast:
        type: integer
      recv_packets:
        type: integer
      sent_bytes:
        type: integer
      sent_carrier:
        type: integer
      sent_colls:
        type: integer
      sent_compressed:
        type: integer
      sent_drop:
        type: integer
      sent_errs:
        type: integer
      sent_fifo:
        type: integer
      sent_packets:
        type: integer
    type: object
  main.SysInfo:
    properties:
      cpus:
        additionalProperties:
          $ref: '#/definitions/main.CPU'
        type: object
      disks:
        items:
          $ref: '#/definitions/main.DiskStat'
        type: array
      filefd:
        $ref: '#/definitions/main.FileFD'
      load_average:
        items:
          type: number
        type: array
      memory:
        $ref: '#/definitions/main.Memory'
      net:
        items:
          $ref: '#/definitions/main.NetworkDevice'
        type: array
    type: object
  main.VMConfig:
    properties:
      config:
        properties:
          auto_destroy:
            type: boolean
          files:
            items:
              properties:
                guest_path:
                  type: string
                raw_value:
                  type: string
              type: object
            type: array
          guest:
            properties:
              cpus:
                type: integer
              memory_mb:
                type: integer
            type: object
          image:
            type: string
          init:
            properties:
              exec:
                items:
                  type: string
                type: array
            type: object
        type: object
    type: object
  main.VMStatus:
    properties:
      ok:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing and controlling Firecracker VMs
  title: Firecracker VM API
  version: "1.0"
paths:
  /create:
    post:
      consumes:
      - application/json
      description: Starts a new Firecracker VM with the provided configuration
      parameters:
      - description: VM Configuration
        in: body
        name: vmConfig
        required: true
        schema:
          $ref: '#/definitions/main.VMConfig'
      produces:
      - application/json
      responses:
        "200":
          description: VM Creation Response
          schema:
            $ref: '#/definitions/main.CreateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Start a new Firecracker VM
  /exec/{machine_id}:
    post:
      consumes:
      - application/json
      description: Executes a command in a running VM
      parameters:
      - description: Machine ID
        in: path
        name: machine_id
        required: true
        type: string
      - description: Command to execute
        in: body
        name: execCmd
        required: true
        schema:
          $ref: '#/definitions/main.ExecCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Command Output
          schema:
            $ref: '#/definitions/main.ExecResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Execute a command in a VM
  /status/{machine_id}:
    get:
      consumes:
      - application/json
      description: Retrieves the status of a running VM
      parameters:
      - description: Machine ID
        in: path
        name: machine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: VM Status
          schema:
            $ref: '#/definitions/main.VMStatus'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get VM Status
  /sys_info/{machine_id}:
    get:
      consumes:
      - application/json
      description: Retrieves system information of a running VM
      parameters:
      - description: Machine ID
        in: path
        name: machine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: System Information
          schema:
            $ref: '#/definitions/main.SysInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get System Information
swagger: "2.0"